+++ src/ext.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "num-bigint")]
+#[cfg_attr(docsrs, doc(cfg(feature = "num-bigint")))]
+mod num_bigint;
+
+#[cfg(feature = "crypto-bigint")]
+#[cfg_attr(docsrs, doc(cfg(feature = "crypto-bigint")))]
+mod crypto_bigint;