+++ src/tests.rs
@@ -3,3 +3,6 @@ mod integers;
 mod rundgren;
 mod simple_value;
 mod value;
+
+#[cfg(all(feature = "num-bigint", feature = "crypto-bigint"))]
+mod bigint_interop;